Create local api alongside firebase api
